Ajuste posicionamento tag (tpAntecip) Tipo de Permissão em relação a antecipação das parcelas conforme manual.
Registro dos transformers faltantes.
