+++ src/main/java/com/fincatto/documentofiscal/transformers/DFRegistryMatcher.java
@@ -1,28 +1,174 @@
 package com.fincatto.documentofiscal.transformers;
 
-import com.fincatto.documentofiscal.DFAmbiente;
-import com.fincatto.documentofiscal.DFModelo;
-import com.fincatto.documentofiscal.DFPais;
-import com.fincatto.documentofiscal.DFUnidadeFederativa;
-import com.fincatto.documentofiscal.cte200.classes.*;
-import com.fincatto.documentofiscal.cte200.transformers.*;
-import com.fincatto.documentofiscal.mdfe3.classes.def.*;
-import com.fincatto.documentofiscal.mdfe3.transformers.*;
-import com.fincatto.documentofiscal.nfe.NFTipoEmissao;
-import com.fincatto.documentofiscal.nfe.transformers.NFTipoEmissaoTransformer;
-import com.fincatto.documentofiscal.nfe310.classes.*;
-import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFIndicadorContribuinteCTe;
-import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFIndicadorContribuinteNFe;
-import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFSituacaoContribuinte;
-import com.fincatto.documentofiscal.nfe310.classes.lote.envio.NFLoteIndicadorProcessamento;
-import com.fincatto.documentofiscal.nfe310.classes.nota.*;
-import com.fincatto.documentofiscal.nfe310.transformers.*;
-import org.simpleframework.xml.transform.RegistryMatcher;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.ZonedDateTime;
+
+import org.simpleframework.xml.transform.RegistryMatcher;
+
+import com.fincatto.documentofiscal.DFAmbiente;
+import com.fincatto.documentofiscal.DFModelo;
+import com.fincatto.documentofiscal.DFPais;
+import com.fincatto.documentofiscal.DFUnidadeFederativa;
+import com.fincatto.documentofiscal.cte200.classes.CTFormaPagamento;
+import com.fincatto.documentofiscal.cte200.classes.CTIdentificadorEmissor;
+import com.fincatto.documentofiscal.cte200.classes.CTModal;
+import com.fincatto.documentofiscal.cte200.classes.CTResponsavelSeguro;
+import com.fincatto.documentofiscal.cte200.classes.CTRetiraMercadoria;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoCte;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoDocumentoOutro;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoDocumentoTransporteAnterior;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoEmissao;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoEntregaData;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoEntregaHorario;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoImpressao;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoServico;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoUnidadeCarga;
+import com.fincatto.documentofiscal.cte200.classes.CTTipoUnidadeTransporte;
+import com.fincatto.documentofiscal.cte200.classes.CTTomadorServico;
+import com.fincatto.documentofiscal.cte200.classes.CTUnidadeMedida;
+import com.fincatto.documentofiscal.cte200.transformers.CTFormaPagamentoTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTIdentificadorEmissorTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTModalTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTResponsavelSeguroTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTRetiraMercadoriaTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoCteTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoDocumentoOutroTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoDocumentoTransporteAnteriorTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoEmissaoTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoEntregaDataTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoEntregaHorarioTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoImpressaoTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoServicoTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoUnidadeCargaTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTipoUnidadeTransporteTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTTomadorServicoTransformer;
+import com.fincatto.documentofiscal.cte200.transformers.CTUnidadeMedidaTransformer;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFModalidadeTransporte;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFProcessoEmissao;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoCargaProdutoPredominante;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoCarroceria;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoEmissao;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoEmitente;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoIndAltoDesempenho;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoInfPag;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoInfPagComp;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoPermissaoAntecipacao;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoProprietario;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoResponsavelSeguro;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoRodado;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoTranportador;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoUnidadeCarga;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoUnidadeTransporte;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFTipoValePedagio;
+import com.fincatto.documentofiscal.mdfe3.classes.def.MDFUnidadeMedidaPesoBrutoCarga;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFModalidadeTransporteTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFProcessoEmissaoTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoCargaProdutoPredominanteTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoCarroceriaTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoEmissaoTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoEmitenteTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoIndAltoDesempenhoTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoInfPagCompTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoInfPagTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoPermissaoAntecipacaoTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoProprietarioTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoResponsavelSeguroTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoRodadoTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoTranportadorTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoUnidadeCargaTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoUnidadeTransporteTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFTipoValePedagioTransformer;
+import com.fincatto.documentofiscal.mdfe3.transformers.MDFUnidadeMedidaPesoBrutoCargaTransformer;
+import com.fincatto.documentofiscal.nfe.NFTipoEmissao;
+import com.fincatto.documentofiscal.nfe.transformers.NFTipoEmissaoTransformer;
+import com.fincatto.documentofiscal.nfe310.classes.NFFinalidade;
+import com.fincatto.documentofiscal.nfe310.classes.NFFormaPagamentoPrazo;
+import com.fincatto.documentofiscal.nfe310.classes.NFModalidadeFrete;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoCombustivelTipo;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoEspecieVeiculo;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoImpostoTributacaoICMS;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemModalidadeBCICMS;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemModalidadeBCICMSST;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemProdutoArmamentoTipo;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemProdutoVeiculoCondicao;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemProdutoVeiculoCondicaoChassi;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemProdutoVeiculoRestricao;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoItemProdutoVeiculoTipoOperacao;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoSituacaoTributariaCOFINS;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoSituacaoTributariaIPI;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoSituacaoTributariaPIS;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoTipoVeiculo;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaInfoVeiculoCor;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaMotivoDesoneracaoICMS;
+import com.fincatto.documentofiscal.nfe310.classes.NFNotaSituacaoOperacionalSimplesNacional;
+import com.fincatto.documentofiscal.nfe310.classes.NFOrigem;
+import com.fincatto.documentofiscal.nfe310.classes.NFOrigemProcesso;
+import com.fincatto.documentofiscal.nfe310.classes.NFProcessoEmissor;
+import com.fincatto.documentofiscal.nfe310.classes.NFProdutoCompoeValorNota;
+import com.fincatto.documentofiscal.nfe310.classes.NFRegimeTributario;
+import com.fincatto.documentofiscal.nfe310.classes.NFTipo;
+import com.fincatto.documentofiscal.nfe310.classes.NFTipoImpressao;
+import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFIndicadorContribuinteCTe;
+import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFIndicadorContribuinteNFe;
+import com.fincatto.documentofiscal.nfe310.classes.cadastro.NFSituacaoContribuinte;
+import com.fincatto.documentofiscal.nfe310.classes.lote.envio.NFLoteIndicadorProcessamento;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFFormaImportacaoIntermediacao;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFFormaPagamentoMoeda;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFIdentificadorLocalDestinoOperacao;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFIndicadorIEDestinatario;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFIndicadorPresencaComprador;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFNotaInfoItemIndicadorExigibilidadeISS;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFNotaInfoItemIndicadorIncentivoFiscal;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFNotaInfoRegimeEspecialTributacao;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFOperacaoConsumidorFinal;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFOperadoraCartao;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFTipoIntegracaoPagamento;
+import com.fincatto.documentofiscal.nfe310.classes.nota.NFViaTransporteInternacional;
+import com.fincatto.documentofiscal.nfe310.transformers.NFFinalidadeTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFFormaImportacaoIntermediacaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFFormaPagamentoMoedaTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFFormaPagamentoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFIdentificadorLocalDestinoOperacaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFIndicadorContribuinteCTTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFIndicadorContribuinteNFeTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFIndicadorIEDestinatarioTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFIndicadorPresencaCompradorTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFLoteIndicadorProcessamentoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFModalidadeFreteTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoCombustivelTipoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoEspecieVeiculoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoImpostoTributacaoICMSTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemIndicadorExigibilidadeISSTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemIndicadorIncentivoFiscalTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemModalidadeBCICMSTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemProdutoArmamentoTipoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemProdutoVeiculoCondicaoChassiTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemProdutoVeiculoCondicaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemProdutoVeiculoRestricaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoItemProdutoVeiculoTipoOperacaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoRegimeEspecialTributacaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoSituacaoTributariaCOFINSTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoSituacaoTributariaIPITransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoSituacaoTributariaPISTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoTipoVeiculoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaInfoVeiculoCorTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaMotivoDesoneracaoICMSTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFNotaSituacaoOperacionalSimplesNacionalTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFOperacaoConsumidorFinalTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFOperadoraCartaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFOrigemProcessoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFOrigemTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFProdutoCompoeValorNotaTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFProgramaEmissorTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFRegimeTributarioTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFSituacaoContribuinteTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFTipoImpressaoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFTipoIntegracaoPagamentoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFTipoTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFViaTransporteInternacionalTransformer;
+import com.fincatto.documentofiscal.nfe310.transformers.NFnotaInfoItemModalidadeBCICMSSTTransformer;
 
 public class DFRegistryMatcher extends RegistryMatcher {
 
@@ -247,5 +393,7 @@ public class DFRegistryMatcher extends RegistryMatcher {
         super.bind(MDFUnidadeMedidaPesoBrutoCarga.class, new MDFUnidadeMedidaPesoBrutoCargaTransformer());
         super.bind(MDFTipoResponsavelSeguro.class, new MDFTipoResponsavelSeguroTransformer());
         super.bind(MDFTipoIndAltoDesempenho.class, new MDFTipoIndAltoDesempenhoTransformer());
+        super.bind(MDFTipoPermissaoAntecipacao.class, new MDFTipoPermissaoAntecipacaoTransformer());
+        super.bind(MDFTipoValePedagio.class, new MDFTipoValePedagioTransformer());
     }
 }
